Dropping activeuuid
Was causing type casting errors that I was unable to figure out after an hour or
so. According to some research we will take a performance penalty by not
storing this data in binary, but it should be fairly small until we reach a huge
number of records. We can always look into changing this in the future if it
becomes a problem.